learned about autherization

+++ example/app/Http/Controllers/JobController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\job;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -24,7 +27,6 @@ class JobController extends Controller
     public function show(Job $job)
     {
         return view('jobs.show', ['job' => $job]);
-
     }
 
     public function store()
@@ -33,43 +35,45 @@ class JobController extends Controller
             'title' => ['required', 'min:3'],
             'salary' => ['required'],
         ]);
-    
+
         Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
-    
+
         return redirect('/jobs');
     }
 
     public function edit(Job $job)
     {
-        return view('jobs.edit', ['job' => $job]);
 
+        return view('jobs.edit', ['job' => $job]);
     }
 
     public function update(Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required'],
         ]);
-    
+
         $job->update([
             'title' => request('title'),
-            'salary'=> request('salary'),
+            'salary' => request('salary'),
         ]);
-    
-        return redirect('/jobs/'. $job->id);
+
+        return redirect('/jobs/' . $job->id);
     }
 
     public function destroy(Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         $job->delete();
 
         return redirect('/jobs');
     }
-
-
 }
